+++ src/AOSongDHT.h
@@ -72,7 +72,7 @@ class AOSongDHT : public Sensor
 {
 public:
     // The constructor - need the power pin, the data pin, and the sensor type
-    AOSongDHT(int powerPin, int dataPin, DHTtype type, int measurementsToAverage = 1);
+    AOSongDHT(int8_t powerPin, int8_t dataPin, DHTtype type, uint8_t measurementsToAverage = 1);
 
     SENSOR_STATUS setup(void) override;
     String getSensorName(void) override;
